OrderController - initial implementation of endpoints. domain/dto package created and moved dto classes inside.
