+++ src/main/java/com/kodilla/ecommercee/controller/OrderController.java
@@ -1,4 +1,34 @@
 package com.kodilla.ecommercee.controller;
 
+import com.kodilla.ecommercee.domain.dto.OrderDto;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@RestController
+@RequestMapping("/v1/ecommercee/order")
 public class OrderController {
+    @RequestMapping(method = RequestMethod.GET, value = "getOrders")
+    public List<OrderDto> getOrders() {
+        return new ArrayList<>();
+    }
+
+    @RequestMapping(method = RequestMethod.POST, value = "createOrder")
+    public void createOrders(@RequestBody OrderDto order) {
+    }
+
+    @RequestMapping(method = RequestMethod.GET, value = "getOrder")
+    public OrderDto getOrder(@RequestParam("orderId") Long orderId) {
+        return new OrderDto(1, 2);
+    }
+
+    @RequestMapping(method = RequestMethod.PUT, value = "updateOrder")
+    public OrderDto updateOrder(@RequestBody OrderDto order) {
+        return order;
+    }
+
+    @RequestMapping(method = RequestMethod.DELETE, value = "deleteOrder")
+    public void deleteOrder(@RequestParam("orderId") Long orderId) {
+    }
 }
