+++ src/main/java/com/kodilla/ecommercee/domain/dto/OrderDto.java
@@ -1,0 +1,13 @@
+package com.kodilla.ecommercee.domain.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+public class OrderDto {
+    private long id;
+    private long productId;
+}
