+++ src/main/java/com/kodilla/ecommercee/domain/dto/ProductDto.java
@@ -1,4 +1,4 @@
-package com.kodilla.ecommercee.domain;
+package com.kodilla.ecommercee.domain.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
